run script for fileitems too, test notes in readme, move tests to proper class
