+++ src/test/scala/nl/knaw/dans/easy/stage/fileitem/FileItemConfSpec.scala
@@ -1,0 +1,9 @@
+package nl.knaw.dans.easy.stage.fileitem
+
+import nl.knaw.dans.easy.stage.AbstractConfSpec
+import org.rogach.scallop.ScallopConf
+
+class FileItemConfSpec extends AbstractConfSpec {
+
+  override def getConf: ScallopConf = new FileItemConf("-".split(" "))
+}
